[chain_spec] move `RelayChain` to chain_spec_helpers.rs
